♻️ Refactor snapshot handling and entity management in ClientApplication

+++ server/include/server/NetworkComponents.hpp
@@ -13,7 +13,7 @@
 namespace server::Component {
 
 struct NetworkId {
-    uint8_t id = 0;
+    uint32_t id = 0;
 };
 
 }  // namespace server::Component
